requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
umap-learn>=0.5.3